split build_shapes rule into offshore, nuts3, country/europe rules
